src/main/java/PosterManager.java
public class PosterManager {
    private String[] films = new String[0];
    private int limit;

    public PosterManager() {
        limit = 10;
    }

    public PosterManager(int limit) {
        this.limit = limit;
    }

    public void add(String film) {
        String[] tmp = new String[films.length + 1];
        for (int i = 0; i < films.length; i++) {
            tmp[i] = films[i];
        }
        tmp[tmp.length - 1] = film;
        films = tmp;
    }

    public String[] findAll() {
        return films;
    }

    public String[] findLast() {
        int resultLength;
        if (films.length > limit) {
            resultLength = limit;
        } else {
            resultLength = films.length;
        }
        String[] reversed = new String[resultLength];
        for (int i = 0; i < resultLength; i++) {
            reversed[i] = films[films.length - 1 - i];
        }
        return reversed;
    }
}

